linux-cec: Fix clippy warnings

+++ linux-cec/src/message.rs
@@ -2178,7 +2178,7 @@ mod test_user_control_pressed {
         instance: Message::UserControlPressed {
             ui_command: operand::UiCommand::Play,
         },
-        bytes: [constants::CEC_OP_UI_CMD_PLAY as u8],
+        bytes: [constants::CEC_OP_UI_CMD_PLAY],
         extra: [Overfull, Empty],
     }
 }
